Clean up test setup.
